requirements.txt
pillow
numpy
matplotlib
pyinstaller